+++ src/field.c
@@ -2,7 +2,7 @@
 
 #include <stdlib.h>
 
-struct field* field_create(const uint8_t width, const uint8_t height)
+struct field* field_create(const uint16_t width, const uint16_t height)
 {
     struct field* field = calloc(1, sizeof(struct field));
 
@@ -33,7 +33,7 @@ void field_set_cur_piece(struct field* field, const enum piece_type piece_type)
 }
 
 
-enum piece_type field_get_cell(const struct field* field, const int8_t cell_x, const int8_t cell_y)
+enum piece_type field_get_cell(const struct field* field, const int16_t cell_x, const int16_t cell_y)
 {
     if (cell_x < 0 || cell_x >= field->width || cell_y < 0 || cell_y >= field->height)
     {
@@ -42,7 +42,7 @@ enum piece_type field_get_cell(const struct field* field, const int8_t cell_x, c
     return field->grid[cell_y * field->width + cell_x];
 }
 
-void field_set_cell(struct field* field, const int8_t cell_x, const int8_t cell_y, const enum piece_type piece_type)
+void field_set_cell(struct field* field, const int16_t cell_x, const int16_t cell_y, const enum piece_type piece_type)
 {
     if (cell_x < 0 || cell_x >= field->width || cell_y < 0 || cell_y >= field->height)
     {
@@ -55,17 +55,17 @@ void field_clear_grid(struct field* field)
 {
     for (uint16_t i = 0; i < field->width * field->height; i++)
     {
-        int8_t cell_x = i % field->width;
-        int8_t cell_y = i / field->width;
+        int16_t cell_x = i % field->width;
+        int16_t cell_y = i / field->width;
         field_set_cell(field, cell_x, cell_y, NONE_TYPE);
     }
 }
 
 
-static bool line_should_clear(const struct field* field, const uint8_t line_index)
+static bool line_should_clear(const struct field* field, const uint16_t line_index)
 {
     bool broken = false;
-    for (int8_t i = 0; i < field->width; i++)
+    for (int16_t i = 0; i < field->width; i++)
     {
         if (field_get_cell(field, i, line_index) == NONE_TYPE)
         {
@@ -75,25 +75,25 @@ static bool line_should_clear(const struct field* field, const uint8_t line_inde
     return !broken;
 }
 
-static void clear_line(struct field* field, const uint8_t line_index)
+static void clear_line(struct field* field, const uint16_t line_index)
 {
-    for (uint8_t i = line_index; i > 0; i--)
+    for (uint16_t i = line_index; i > 0; i--)
     {
-        for (uint8_t j = 0; j < field->width; j++)
+        for (uint16_t j = 0; j < field->width; j++)
         {
             field_set_cell(field, j, i, field_get_cell(field, j, i - 1));
         }
     }
-    for (uint8_t i = 0; i < field->width; i++)
+    for (uint16_t i = 0; i < field->width; i++)
     {
         field_set_cell(field, i, 0, NONE_TYPE);
     }
 }
 
-uint8_t field_clear_lines(struct field* field)
+uint16_t field_clear_lines(struct field* field)
 {
-    uint8_t lines_cleared = 0;
-    for (uint8_t i = 0; i < field->height; i++)
+    uint16_t lines_cleared = 0;
+    for (uint16_t i = 0; i < field->height; i++)
     {
         if (line_should_clear(field, i))
         {
@@ -105,18 +105,18 @@ uint8_t field_clear_lines(struct field* field)
 }
 
 
-uint8_t field_get_draw_width(const struct field* field)
+uint16_t field_get_draw_width(const struct field* field)
 {
     return field->width * 2 + 2;
 }
 
-uint8_t field_get_draw_height(const struct field* field)
+uint16_t field_get_draw_height(const struct field* field)
 {
     return field->height + 2;
 }
 
 
-bool field_cur_piece_collides(const struct field* field, const int8_t dx, const int8_t dy, const int8_t rotation)
+bool field_cur_piece_collides(const struct field* field, const int16_t dx, const int16_t dy, const int16_t rotation)
 {
     for (int32_t i = 0; i < PIECE_NUM_SQUARES; i++)
     {
@@ -138,19 +138,19 @@ bool field_cur_piece_collides(const struct field* field, const int8_t dx, const 
 
 void field_lock_cur_piece(struct field* field)
 {
-    for (uint8_t i = 0; i < PIECE_NUM_SQUARES; i++)
+    for (uint16_t i = 0; i < PIECE_NUM_SQUARES; i++)
     {
-        int8_t cell_x = field->pos_x + field->cur_piece->coordinates[field->cur_piece->rotation][i][0];
-        int8_t cell_y = field->pos_y + field->cur_piece->coordinates[field->cur_piece->rotation][i][1];
+        int16_t cell_x = field->pos_x + field->cur_piece->coordinates[field->cur_piece->rotation][i][0];
+        int16_t cell_y = field->pos_y + field->cur_piece->coordinates[field->cur_piece->rotation][i][1];
         field_set_cell(field, cell_x, cell_y, field->cur_piece->type);
     }
 }
 
 bool field_should_lose(struct field* field)
 {
-    for (int8_t i = 0; i < PIECE_NUM_SQUARES; i++)
+    for (int16_t i = 0; i < PIECE_NUM_SQUARES; i++)
     {
-        int8_t cell_y = field->pos_y + field->cur_piece->coordinates[field->cur_piece->rotation][i][1];
+        int16_t cell_y = field->pos_y + field->cur_piece->coordinates[field->cur_piece->rotation][i][1];
         if (cell_y < 0)
         {
             return true;
@@ -160,7 +160,7 @@ bool field_should_lose(struct field* field)
 }
 
 
-bool field_move_cur_piece(struct field* field, const int8_t dx, const int8_t dy, bool do_collision_check)
+bool field_move_cur_piece(struct field* field, const int16_t dx, const int16_t dy, bool do_collision_check)
 {
     if (do_collision_check && field_cur_piece_collides(field, dx, dy, field->cur_piece->rotation))
     {
@@ -171,9 +171,9 @@ bool field_move_cur_piece(struct field* field, const int8_t dx, const int8_t dy,
     return true;
 }
 
-uint8_t field_get_lowest_height(const struct field* field)
+uint16_t field_get_lowest_height(const struct field* field)
 {
-    for (int8_t i = field->pos_y + 1; i < field->height; i++)
+    for (int16_t i = field->pos_y + 1; i < field->height; i++)
     {
         if (field_cur_piece_collides(field, 0, i - field->pos_y, field->cur_piece->rotation))
         {
@@ -183,7 +183,7 @@ uint8_t field_get_lowest_height(const struct field* field)
     return 0;
 }
 
-int8_t field_get_highest_square_height(const struct field* field)
+int16_t field_get_highest_square_height(const struct field* field)
 {
     for (int16_t i = 0; i < field->width * field->height; i++)
     {
@@ -206,9 +206,9 @@ void field_soft_drop_cur_piece(struct field* field)
     field_move_cur_piece(field, 0, field_get_lowest_height(field) - field->pos_y, false);
 }
 
-bool field_rotate_cur_piece(struct field* field, const int8_t direction)
+bool field_rotate_cur_piece(struct field* field, const int16_t direction)
 {
-    uint8_t attempted_rotation = piece_get_next_rotation(field->cur_piece, direction);
+    uint16_t attempted_rotation = piece_get_next_rotation(field->cur_piece, direction);
     if (!field_cur_piece_collides(field, 0, 0, attempted_rotation))
     {
         piece_rotate(field->cur_piece, direction);
@@ -217,8 +217,8 @@ bool field_rotate_cur_piece(struct field* field, const int8_t direction)
 
     for (int i = 0; i < PIECE_NUM_COLLISION_CHECKS; i++)
     {
-        int8_t check_x = field->cur_piece->collision_checks[field->cur_piece->rotation][direction == 1][i][0];
-        int8_t check_y = field->cur_piece->collision_checks[field->cur_piece->rotation][direction == 1][i][1];
+        int16_t check_x = field->cur_piece->collision_checks[field->cur_piece->rotation][direction == 1][i][0];
+        int16_t check_y = field->cur_piece->collision_checks[field->cur_piece->rotation][direction == 1][i][1];
 
         if (!field_cur_piece_collides(field, check_x, check_y, attempted_rotation))
         {
