+++ src/piece-data.c
@@ -1,6 +1,6 @@
 #include <stdint.h>
 
-static const uint8_t piece_data_collision_coords[7][4][4][2] = 
+static const uint16_t piece_data_collision_coords[7][4][4][2] = 
 {
     // I PIECE:
     {
@@ -223,7 +223,7 @@ static const uint8_t piece_data_collision_coords[7][4][4][2] =
     } 
 };
 
-static const int8_t piece_data_collision_checks[7][4][2][4][2] = 
+static const int16_t piece_data_collision_checks[7][4][2][4][2] = 
 {
     // I PIECE:
     {
