+++ src/piece.h
@@ -27,14 +27,14 @@ enum piece_type
 struct piece
 {
     enum piece_type type;
-    int8_t rotation;
-    uint8_t coordinates[PIECE_NUM_ROTATIONS][PIECE_NUM_TYPES][2];
-    int8_t collision_checks[PIECE_NUM_ROTATIONS][2][PIECE_NUM_COLLISION_CHECKS][2];
+    int16_t rotation;
+    uint16_t coordinates[PIECE_NUM_ROTATIONS][PIECE_NUM_TYPES][2];
+    int16_t collision_checks[PIECE_NUM_ROTATIONS][2][PIECE_NUM_COLLISION_CHECKS][2];
 };
 
 struct piece piece_create(const enum piece_type type);
 
-void piece_rotate(struct piece* tetromino, const int8_t direction);
-uint8_t piece_get_next_rotation(const struct piece* piece, const int8_t direction);
+void piece_rotate(struct piece* tetromino, const int16_t direction);
+uint16_t piece_get_next_rotation(const struct piece* piece, const int16_t direction);
 
 #endif
